Replace CPS Debug feature with PAL Debugger
- Remove `CpsDebugScreen` and `CpsDebugViewModel` and replace with `PalDebugScreen` and `PalDebugViewModel`.
- Implement `PalDebugScreen` to list available PAL files and render selected palettes on a `Canvas`.
- Create `ResourceRepository` to handle reading raw bytes and listing resource files.
- Update `PalRepository` to parse raw PAL data, converting 6-bit color values to 8-bit.
- Update `RGB` data model to use `UByte` and `Pal` to use `ImmutableList`.
- Register `ResourceRepository` and `PalDebugViewModel` in Koin modules.
- Add `kermit` and `kotlinx-collections-immutable` dependencies to version catalog and build file.

+++ composeApp/src/commonMain/kotlin/pl/pelotasplus/eyeofbeholder/data/model/RGB.kt
@@ -1,7 +1,7 @@
 package pl.pelotasplus.eyeofbeholder.data.model
 
 data class RGB(
-    val red: Int,
-    val green: Int,
-    val blue: Int,
+    val red: UByte,
+    val green: UByte,
+    val blue: UByte,
 )
